change vmstat 1 1 to vmstat 1 2

+++ python/instrallpython3.6.4.py
@@ -2,6 +2,10 @@
 #coding:utf-8
 import subprocess,os
 
+if os.path.exists('/usr/bin/python3'):
+    print('python3 is exist already.')
+    exit(1)
+
 subprocess.call('yum install gcc zlib zlib-devel libffi-devel -y',shell=True)
 os.chdir('/tmp')
 subprocess.call('curl -fsSL https://www.python.org/ftp/python/3.6.5/Python-3.6.5.tar.xz > python3.6.5.tar.xz',shell=True)
@@ -11,4 +15,4 @@ subprocess.call('./configure --prefix=/usr/local/python3',shell=True)
 subprocess.call('make',shell=True)
 subprocess.call('make install',shell=True)
 subprocess.call('ln -s /usr/local/python3/bin/python3 /usr/bin/',shell=True)
-subprocess.call('python3 --version',shell=True)
+subprocess.call('python3 --version',shell=True
